fix(views): make preview-lock button only show up in preview
