+++ packages/dendron-plugin-views/src/components/DendronApp.tsx
@@ -20,16 +20,13 @@ import {
   Provider,
   setLogLevel,
 } from "@dendronhq/common-frontend";
-import { Layout, Button } from "antd";
-import LockFilled from "@ant-design/icons/lib/icons/LockFilled";
-import UnlockOutlined from "@ant-design/icons/lib/icons/UnlockOutlined";
+import { Layout } from "antd";
 import _ from "lodash";
 import React from "react";
 import { useWorkspaceProps } from "../hooks";
 import "../styles/scss/main.scss";
 import { DendronComponent } from "../types";
 import { postVSCodeMessage, useVSCodeMessage } from "../utils/vscode";
-import type { SyntheticEvent } from "react";
 
 const { Content } = Layout;
 
@@ -182,40 +179,7 @@ function DendronVSCodeApp({ Component }: { Component: DendronComponent }) {
     }
   });
 
-  const isLocked = props.ide.isLocked;
-
-  const handleLock = (event: SyntheticEvent<HTMLElement>) => {
-    if (!(event.currentTarget instanceof HTMLElement)) {
-      return;
-    }
-
-    postVSCodeMessage({
-      type: isLocked
-        ? NoteViewMessageEnum.onUnlock
-        : NoteViewMessageEnum.onLock,
-      data: {
-        id: props.ide.noteActive?.id,
-      },
-      source: DMessageSource.webClient,
-    });
-  };
-
-  return (
-    <>
-      <Component {...props} />
-      <Button
-        shape="circle"
-        icon={isLocked ? <LockFilled /> : <UnlockOutlined />}
-        onClick={handleLock}
-        style={{
-          position: "absolute",
-          top: 33,
-          right: 33,
-          opacity: isLocked ? 1 : 0.3,
-        }}
-      />
-    </>
-  );
+  return <Component {...props} />;
 }
 
 export type DendronAppProps = {
