+++ packages/dendron-plugin-views/src/components/DendronNotePreview.tsx
@@ -11,6 +11,9 @@ import {
 } from "@dendronhq/common-frontend";
 import _ from "lodash";
 import mermaid from "mermaid";
+import { Button } from "antd";
+import LockFilled from "@ant-design/icons/lib/icons/LockFilled";
+import UnlockOutlined from "@ant-design/icons/lib/icons/UnlockOutlined";
 import React from "react";
 import {
   useCurrentTheme,
@@ -20,6 +23,7 @@ import {
 } from "../hooks";
 import { DendronComponent } from "../types";
 import { postVSCodeMessage } from "../utils/vscode";
+import type { SyntheticEvent } from "react";
 
 function isHTMLAnchorElement(element: Element): element is HTMLAnchorElement {
   return element.nodeName === "A";
@@ -122,7 +126,39 @@ const DendronNotePreview: DendronComponent = (props) => {
     return <div>Loading...</div>;
   }
 
-  return <DendronNote noteContent={noteRenderedBody} config={config} />;
+  const isLocked = props.ide.isLocked;
+
+  const handleLock = (event: SyntheticEvent<HTMLElement>) => {
+    if (!(event.currentTarget instanceof HTMLElement)) {
+      return;
+    }
+    postVSCodeMessage({
+      type: isLocked
+        ? NoteViewMessageEnum.onUnlock
+        : NoteViewMessageEnum.onLock,
+      data: {
+        id: props.ide.noteActive?.id,
+      },
+      source: DMessageSource.webClient,
+    });
+  };
+
+  return (
+    <>
+      <DendronNote noteContent={noteRenderedBody} config={config} />;
+      <Button
+        shape="circle"
+        icon={isLocked ? <LockFilled /> : <UnlockOutlined />}
+        onClick={handleLock}
+        style={{
+          position: "absolute",
+          top: 33,
+          right: 33,
+          opacity: isLocked ? 1 : 0.3,
+        }}
+      />
+    </>
+  );
 };
 
 export default DendronNotePreview;
